+++ spec/adapterStub.js
@@ -11,9 +11,10 @@ var mocks = {
 	"/analytics/api/board/101/lane/307": {},
 	"/analytics/api/nonstop/package?project=one&build=1&version=0.1.0": { packages: [] },
 	"/analytics/api/nonstop/project?owner=me&build=1&version=0.1.0": { projects: [] },
-	"http://localhost:8088/analytics/api/nonstop/upload": {},
+	"http://localhost:8088/analytics/api/nonstop/upload": "",
 	"http://localhost:8088/analytics/api/elevated/gimme": "User lacks sufficient permissions",
-	"http://localhost:8088/analytics/api/this/204": ""
+	"http://localhost:8088/analytics/api/this/204": "",
+	"http://localhost:8088/analytics/bad/request": require( "./mockResponses/badRequest.json" )
 };
 var adapter = function( resultsArray ) {
 	return function( link, data ) {