+++ spec/requestStub.js
@@ -7,10 +7,12 @@ module.exports = function( adapterStub ) {
 		adapterStub( { href: opts.url }, opts )
 			.then( function( res ) {
 				var body = "";
+				var status = res.statusCode ? res.statusCode : 200;
+				delete res.statusCode;
 				if ( res ) {
 					body = _.isString( res ) ? res : JSON.stringify( res );
 				}
-				cb( null, null, body );
+				cb( undefined, { statusCode: status }, body );
 			} )
 			.catch( function( err ) {
 				cb( err, null, null );